refactor(PlaceMap): 비대해진 ViewModel 3개의 Handler로 분리
비대해진 `PlaceMapViewModel`의 책임을 분산하기 위해 Action 처리 로직을 기능별 핸들러로 위임하고, 관련 패키지 구조를 `intent` 중심으로 재구성했습니다.

- **Action 및 Handler 세분화:**
    - 단일 `PlaceMapAction`을 `SelectAction`(장소 선택 및 상호작용), `FilterAction`(필터링 및 데이터 로드), `MapEventAction`(지도 제어 이벤트)으로 분리했습니다.
    - 각 Action을 전담하여 처리하는 `SelectActionHandler`, `FilterActionHandler`, `MapEventActionHandler`를 도입하여 ViewModel의 로직을 분리했습니다.

- **패키지 구조 재편:**
    - `viewmodel` 패키지에 혼재되어 있던 클래스들을 `intent` 패키지 하위의 `action`, `event`, `state`로 이동하여 구조를 명확히 했습니다.
    - `PlaceMapUiState`, `LoadState`, `ListLoadState` 등의 상태 클래스와 `MapControlEvent` 등의 이벤트 클래스가 해당 패키지로 이동되었습니다.

- **UI 계층 수정:**
    - `PlaceMapScreen` 및 `PlaceMapFragment`에서 ViewModel로 이벤트를 전달할 때, 기존의 포괄적인 `PlaceMapAction` 대신 구체화된 Action(`SelectAction.OnPlaceClick` 등)을 사용하도록 호출부를 수정했습니다.
