+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/component/PlaceMapScreen.kt
@@ -11,10 +11,13 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.unit.dp
-import com.daedan.festabook.presentation.placeMap.model.LoadState
-import com.daedan.festabook.presentation.placeMap.viewmodel.MapDelegate
-import com.daedan.festabook.presentation.placeMap.viewmodel.PlaceMapAction
-import com.daedan.festabook.presentation.placeMap.viewmodel.PlaceMapUiState
+import com.daedan.festabook.presentation.placeMap.intent.action.FilterAction
+import com.daedan.festabook.presentation.placeMap.intent.action.MapEventAction
+import com.daedan.festabook.presentation.placeMap.intent.action.PlaceMapAction
+import com.daedan.festabook.presentation.placeMap.intent.action.SelectAction
+import com.daedan.festabook.presentation.placeMap.intent.state.LoadState
+import com.daedan.festabook.presentation.placeMap.intent.state.MapDelegate
+import com.daedan.festabook.presentation.placeMap.intent.state.PlaceMapUiState
 import com.daedan.festabook.presentation.theme.FestabookColor
 import com.daedan.festabook.presentation.theme.festabookSpacing
 
@@ -29,8 +32,8 @@ fun PlaceMapScreen(
     NaverMapContent(
         modifier = modifier.fillMaxSize(),
         mapDelegate = mapDelegate,
-        onMapReady = { onAction(PlaceMapAction.OnMapReady) },
-        onMapDrag = { onAction(PlaceMapAction.OnMapDrag) },
+        onMapReady = { onAction(MapEventAction.OnMapReady) },
+        onMapDrag = { onAction(MapEventAction.OnMapDrag) },
     ) { naverMap ->
         Column(
             modifier = Modifier.wrapContentSize(),
@@ -39,7 +42,7 @@ fun PlaceMapScreen(
                 timeTagsState = uiState.timeTags,
                 selectedTimeTagState = uiState.selectedTimeTag,
                 onTimeTagClick = { timeTag ->
-                    onAction(PlaceMapAction.OnTimeTagClick(timeTag))
+                    onAction(SelectAction.OnTimeTagClick(timeTag))
                 },
                 modifier =
                     Modifier
@@ -50,8 +53,8 @@ fun PlaceMapScreen(
             PlaceCategoryScreen(
                 initialCategories = uiState.initialCategories,
                 selectedCategories = uiState.selectedCategories,
-                onCategoryClick = { onAction(PlaceMapAction.OnCategoryClick(it)) },
-                onDisplayAllClick = { onAction(PlaceMapAction.OnCategoryClick(it)) },
+                onCategoryClick = { onAction(FilterAction.OnCategoryClick(it)) },
+                onDisplayAllClick = { onAction(FilterAction.OnCategoryClick(it)) },
             )
 
             Box(
@@ -75,12 +78,12 @@ fun PlaceMapScreen(
                         ),
                     placesUiState = uiState.places,
                     map = naverMap,
-                    onPlaceClick = { onAction(PlaceMapAction.OnPlaceClick(it.id)) },
+                    onPlaceClick = { onAction(SelectAction.OnPlaceClick(it.id)) },
                     bottomSheetState = bottomSheetState,
                     isExceededMaxLength = uiState.isExceededMaxLength,
-                    onPlaceLoadFinish = { onAction(PlaceMapAction.OnPlaceLoadFinish(it)) },
-                    onPlaceLoad = { onAction(PlaceMapAction.OnPlaceLoad) },
-                    onBackToInitialPositionClick = { onAction(PlaceMapAction.OnBackToInitialPositionClick) },
+                    onPlaceLoadFinish = { onAction(MapEventAction.OnPlaceLoadFinish(it)) },
+                    onPlaceLoad = { onAction(FilterAction.OnPlaceLoad) },
+                    onBackToInitialPositionClick = { onAction(MapEventAction.OnBackToInitialPositionClick) },
                 )
 
                 if (uiState.isPlacePreviewVisible) {
@@ -94,8 +97,8 @@ fun PlaceMapScreen(
                                 ),
                         selectedPlace = uiState.selectedPlace,
                         visible = true,
-                        onClick = { onAction(PlaceMapAction.OnPlacePreviewClick(it)) },
-                        onBackPress = { onAction(PlaceMapAction.OnBackPress) },
+                        onClick = { onAction(SelectAction.OnPlacePreviewClick(it)) },
+                        onBackPress = { onAction(SelectAction.OnBackPress) },
                     )
                 }
 
@@ -110,7 +113,7 @@ fun PlaceMapScreen(
                                 ),
                         selectedPlace = uiState.selectedPlace,
                         visible = true,
-                        onBackPress = { onAction(PlaceMapAction.OnBackPress) },
+                        onBackPress = { onAction(SelectAction.OnBackPress) },
                     )
                 }
             }
