+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/intent/action/FilterAction.kt
@@ -1,0 +1,11 @@
+package com.daedan.festabook.presentation.placeMap.intent.action
+
+import com.daedan.festabook.presentation.placeMap.model.PlaceCategoryUiModel
+
+sealed interface FilterAction : PlaceMapAction {
+    data class OnCategoryClick(
+        val categories: Set<PlaceCategoryUiModel>,
+    ) : FilterAction
+
+    data object OnPlaceLoad : FilterAction
+}
