+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/intent/action/PlaceMapAction.kt
@@ -1,0 +1,3 @@
+package com.daedan.festabook.presentation.placeMap.intent.action
+
+sealed interface PlaceMapAction
