+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/intent/state/MapDelegate.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeMap.viewmodel
+package com.daedan.festabook.presentation.placeMap.intent.state
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
