+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/intent/state/PlaceMapUiState.kt
@@ -1,14 +1,11 @@
-package com.daedan.festabook.presentation.placeMap.viewmodel
+package com.daedan.festabook.presentation.placeMap.intent.state
 
 import com.daedan.festabook.domain.model.TimeTag
 import com.daedan.festabook.presentation.placeDetail.model.PlaceDetailUiModel
 import com.daedan.festabook.presentation.placeMap.model.InitialMapSettingUiModel
-import com.daedan.festabook.presentation.placeMap.model.ListLoadState
-import com.daedan.festabook.presentation.placeMap.model.LoadState
 import com.daedan.festabook.presentation.placeMap.model.PlaceCategoryUiModel
 import com.daedan.festabook.presentation.placeMap.model.PlaceCoordinateUiModel
 import com.daedan.festabook.presentation.placeMap.model.PlaceUiModel
-import com.daedan.festabook.presentation.placeMap.model.isSecondary
 
 data class PlaceMapUiState(
     val initialMapSetting: LoadState<InitialMapSettingUiModel> = LoadState.Loading,
