--- app/src/main/java/com/daedan/festabook/presentation/placeMap/viewmodel/PlaceMapActionHandler.kt
@@ -1,3 +1,0 @@
-package com.daedan.festabook.presentation.placeMap.viewmodel
-
-class PlaceMapActionHandler
